requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
open-clip-torch==2.24.0
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
numpy==1.26.4
onnxruntime==1.18.0
scipy==1.13.0
